+++ demos/demo-supabase-auth/src/middlewares/errorHandler.ts
@@ -1,22 +1,32 @@
 import { MoralisError } from '@moralisweb3/core';
 import { NextFunction, Request, Response } from 'express';
 
-export function errorHandler(error: Error | MoralisError, req: Request, res: Response, next: NextFunction) {
-  if (error instanceof MoralisError) {
-    const errorResponse = error.details?.response;
-    const errorResponseData =
-      typeof errorResponse == 'object' ? (error.details?.response as Record<string, any>).data : null;
+const makeMoralisErrorMessage = (error: MoralisError) => {
+  let message = error.message || 'Unknown error';
 
-    const status = typeof error.details?.status === 'number' ? error.details?.status : 500;
-    let errorMessage = error.message || 'Unknown error';
+  const errorResponse = error.details?.response;
 
-    if (errorResponseData) {
-      if (errorResponseData && errorResponseData?.message) {
-        errorMessage = `${errorResponseData?.name ? `${errorResponseData.name}: ` : ''}${errorResponseData.message}`;
-      } else {
-        errorMessage = 'Unknown error';
-      }
+  const errorResponseData =
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    typeof errorResponse === 'object' ? (error.details?.response as Record<string, any>).data : null;
+
+  if (errorResponseData) {
+    // Handle MoralisError
+    if (errorResponseData && errorResponseData?.message) {
+      message = `${errorResponseData?.name ? `${errorResponseData.name}: ` : ''}${errorResponseData.message}`;
+    } else if (errorResponseData.error) {
+      // Handle ParseError
+      message = errorResponseData.error;
     }
+  }
+
+  return message;
+};
+
+export function errorHandler(error: Error | MoralisError, req: Request, res: Response, _next: NextFunction) {
+  if (error instanceof MoralisError) {
+    const status = typeof error.details?.status === 'number' ? error.details?.status : 500;
+    const errorMessage = makeMoralisErrorMessage(error);
 
     res.status(status).json({ error: errorMessage });
   } else {