+++ packages/evmUtils/src/dataTypes/Erc20/Erc20.ts
@@ -17,7 +17,7 @@ export class Erc20Token implements MoralisDataObject {
   }
 
   static parse = (value: Erc20Input): Erc20Data => ({
-    decimals: maybe(value.decimals, (value) => +value),
+    decimals: +value.decimals,
     name: value.name,
     symbol: value.symbol,
     contractAddress: EvmAddress.create(value.contractAddress),